+++ BankAccountBusiness/IBank.cs
@@ -5,8 +5,8 @@ namespace BankAccount.Business
     public interface IBank
     {
         int CreateAccount(string firstName, string name);
-        bool Deposit(int accountId, double value);
-        bool Withdraw(int accountId, double value);
+        bool Deposit(int accountId, decimal value);
+        bool Withdraw(int accountId, decimal value);
         Statement GetStatement(int accountId);
         Account GetAccount(int accountId);
     }
